Create getDaysInMonth
Write a method isLeapYear with a parameter of type int named year. 

The parameter needs to be greater than or equal to 1 and less than or equal to 9999.

If the parameter is not in that range return false. 

Otherwise, if it is in the valid range, calculate if the year is a leap year and return true if it is, otherwise return false. 

A year is a leap year if it is divisible by 4 but not by 100, or it is divisible by 400.

********AND*********

Write another method getDaysInMonth with two parameters month and year.  ​Both of type int.

If parameter month is &lt; 1 or &gt; 12 return -1. ​
If parameter year is &lt; 1 or &gt; 9999 then return -1.

This method needs to return the number of days in the month. Be careful about leap years they have 29 days in month 2 (February).

You should check if the year is a leap year using the method isLeapYear described above.